Move "ensure columns" handling to AbstractTableInfo for reliability and wider use
